move existing test to separate file
